+++ tests/Helper/ObjectParserTest.php
@@ -178,4 +178,18 @@ final class ObjectParserTest extends TestCase
         $this->assertNotSame($rules, $parser->getRules());
         $this->assertArrayNotHasKey($cacheKey, $cacheProperty->getValue());
     }
+
+    public function testUninitializedProperty(): void
+    {
+        $parser = new ObjectParser(new class () {
+            public int $a = 4;
+            public int $uninitialized;
+        });
+
+        $this->assertNull($parser->getPropertyValue('not-existing'));
+        $this->assertSame(4, $parser->getPropertyValue('a'));
+        $this->assertNull($parser->getPropertyValue('uninitialized'));
+
+        $this->assertSame(['a' => 4], $parser->getData());;
+    }
 }