+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.5.1 under development
 
 - Bug #776: Take `PropertyTranslatorProviderInterface` into account during context validation (@vjik)
+- Enh #???: Default Intl locale in tests (@batyrmastyr)
+- Enh #???: Support validation of uninitialized properties (@batyrmastyr)
 
 ## 2.5.0 July 19, 2025